Create read-me.md
# Apps Script: Footage Entry Tool for Spreadsheet

This is a sidebar form designed specifically to partially-automate and generally simplify the footage logging process for the production of youtube series [Modern Rogue](https://youtube.com/modernrogue). We've got a lot of data for every batch of footage, and the easier I can make that entry process, the less likely data will get logged incorrectly.