conditional function and formatting
The passToTab2 function is now conditional, so it only populates the release order tab with episodes that are completed. The release tab needs these to start scheduling releases, but ads, patreon bonuses, and incomplete videos do not need to be scheduled in the same manner.

Also some spaces were added in between function arguments, and arguments in function calls are exposed to more easily see what variables it's passing to them.