added loading animation
Spinner animation appears next to "Type of Video" to indicate when data is still loading. The animation is removed upon spreadsheet data loading in, which usually takes maybe 2 seconds or so. It's best practice to not change type of video until this disappears. The form can still be used while loading, but auto-completes for sponsor video types might not load in appropriately if switched to too fast.