improved sustainability
The offsets that were used to prevent the data requests from pulling all rows of data (nearly a thousand) would need continual updates (since they were a hard coded number) to maintain a large enough sample to be useful but small enough to not bog down GAS. There seems to be a threshold of a hundred or two samples that scale load times from a few seconds to a few hundred. However, now it is anchored relative to the amount of rows, meaning it should only need to be updated if somebody is experiencing persistent performance issues, but a sample size of 250 rows is a nice balance in my experience.

I've also globalized the newRow variable since it's used in almost every function here.