status colors for crew role buttons
Made it so that whenever any role data is filled in for a particular crew member, their respective button changes color. It then changed back to default if the data is removed (either manually, or by "reset.") This allows the user to easily and visually identify which crew roles are filled in even when all fields are collapsed.