change color theme, crew buttons re-arranged, checkbox toggle hotkey, auto-scroll
Integrated style to change the color theme.

Crew buttons are now 2 rows and smaller in size. This just reduces the vertical height of the form slightly.

Memory card checkboxes now have a hotkey to toggle a series all at once by holding the ctrl key while selecting any given card.

The form scrolls down automatically upon clicking any crew button. This is so that when expanding the roles, the options no longer extend beyond the bottom edge of the form.