data submit fix and optimizations
Fixing some things I unknowingly broke a few commits ago. getLastRow as a global variable returns the number of the last-most entry, whereas when it was a local variable, it returned the new entry row. So I've added +1 to rectify this while keeping it a global variable.

I also noticed data submission was far slower than it used to be. I think I tracked a fair portion of the slowdown to cell formatting. I optimized this by setting a default font size and removing some redundant code.