status indicator and date fix
Date was improperly formatted on a few days and months, causing the date field to not default to a value during those dates, nor allowing data submission on those dates. This should now be fixed.

The submit button now changes text and a spinning indicator appears next to it while the script works on changing cells. It reverts back to default upon completion of the script.